+++ src/js/graphSetup.js
@@ -2,7 +2,7 @@ var setupOrgBudget = function ($org) {
   var options = {
     title: 'Recipient organisation budget',
     el: 'recipient-org-budget',
-    hasStatuses: true,
+    defaultStatus: 'Indicative',
     filter: {
       name: 'organisation',
       el: 'recipient-org',
@@ -21,7 +21,7 @@ var setupRegionBudget = function ($org, codelists) {
   var options = {
     title: 'Recipient region budget',
     el: 'recipient-region-budget',
-    hasStatuses: true,
+    defaultStatus: 'Indicative',
     filter: {
       name: 'region',
       el: 'recipient-region',
@@ -41,7 +41,7 @@ var setupCountryBudget = function ($org, codelists) {
   var options = {
     title: 'Recipient country budget',
     el: 'recipient-country-budget',
-    hasStatuses: true,
+    defaultStatus: 'Indicative',
     filter: {
       name: 'country',
       el: 'recipient-country',
@@ -61,7 +61,7 @@ var setupTotalBudget = function ($org) {
   var options = {
     title: 'Total budget',
     el: 'total-budget',
-    hasStatuses: true,
+    defaultStatus: 'Indicative',
     filter: null,
     breakdown: {
       name: 'budget line',
@@ -75,7 +75,7 @@ var setupTotalExpenditure = function ($org) {
   var options = {
     title: 'Total expenditure',
     el: 'total-expenditure',
-    hasStatuses: false,
+    defaultStatus: 'Expenditure',
     filter: null,
     breakdown: {
       name: 'expense line',